+++ CHANGELOG.rst
@@ -23,6 +23,10 @@ Unreleased
 * Added ``compas_fab.robots.Configuration.iter_differences``
 * Added ``compas_fab.robots.Configuration.max_difference``
 * Added ``compas_fab.robots.Configuration.close_to``
+* Added ``compas_fab.robots.Configuration.merge``
+* Added ``compas_fab.robots.JointTrajectoryPoint.merge``
+* Added ``compas_fab.robots.Semantics.group_states``
+* Added ``compas_fab.robots.Robot.get_configuration_from_group_state``
 
 **Changed**
 
@@ -33,6 +37,7 @@ Unreleased
 * Parameter ``backend`` of forward kinematics has been renamed to ``solver``
 * The signatures of all kinematics, motion planning and planning scene management methods have been homogenized across backend clients and within ``Robot``
 * All examples have been updated to reflect these changes
+* The installer to Rhino has been unified with COMPAS core. Now running ``python -m compas_rhino.install`` will also detect and install COMPAS FAB and its dependencies.
 * Renamed all ``RobotArtist`` implementations to ``RobotModelArtist`` to reflect
   the fact they depend on ``compas.robots.RobotModel``.
 * Renamed  ``compas_fab.robots.Robot.from_tool0_to_attached_tool`` to ``compas_fab.robots.Robot.from_t0cf_to_tcf``